Add tests and move validate_that_golem_messages_are_signed_with_key to validation.py
