+++ concent_api/core/tests/test_unit_message_handlers.py
@@ -2,12 +2,16 @@ from unittest import TestCase
 
 from django.conf import settings
 from django.test import override_settings
+from golem_messages import message
+
+from core.exceptions import GolemMessageValidationError
 from core.message_handlers import are_items_unique
 from core.message_handlers import store_subtask
 from core.models import Subtask
 from core.tests.utils import ConcentIntegrationTestCase
 from core.tests.utils import parse_iso_date_to_timestamp
 from core.utils import hex_to_bytes_convert
+from core.validation import validate_reject_report_computed_task
 
 
 class TestMessageHandlers(TestCase):
@@ -64,3 +68,29 @@ class TestMessagesStored(ConcentIntegrationTestCase):
             parse_iso_date_to_timestamp(subtask.report_computed_task.timestamp.isoformat()),
             parse_iso_date_to_timestamp(self.report_computed_task_timestamp)
         )
+
+
+class TestValidateRejectReportComputedTask(ConcentIntegrationTestCase):
+
+    def test_that_validation_passes_if_correct_message_given(self):
+        correct_reject_reported_computed_task = self._get_deserialized_reject_report_computed_task(
+            reason=message.RejectReportComputedTask.REASON.SubtaskTimeLimitExceeded,
+            task_to_compute=self._get_deserialized_task_to_compute()
+        )
+        validate_reject_report_computed_task(correct_reject_reported_computed_task)
+
+    def test_that_validation_raises_error_if_invalid_reason_instance_given(self):
+        incorrect_reject_reported_computed_task = self._get_deserialized_reject_report_computed_task(
+            reason=message.tasks.CannotAssignTask.REASON.NotMyTask,
+            task_to_compute=self._get_deserialized_task_to_compute()
+        )
+        with self.assertRaises(GolemMessageValidationError):
+            validate_reject_report_computed_task(incorrect_reject_reported_computed_task)
+
+    def test_that_validation_raises_error_if_reason_is_none(self):
+        incorrect_reject_reported_computed_task = self._get_deserialized_reject_report_computed_task(
+            reason=None,
+            task_to_compute=self._get_deserialized_task_to_compute()
+        )
+        with self.assertRaises(GolemMessageValidationError):
+            validate_reject_report_computed_task(incorrect_reject_reported_computed_task)
